Fix EmoStroop ignoring parameters and not working
EmoStroop ignored parameters passed with starting Intent and therefore
did not set up the test properly. An error message was shown and start
button skipped directly to final result display. The fault was the use
of the method fillStringArray() from base class CognitiveActivity into
which a null was passed causing a NullPointerException. Passing arrays
would have solved the Exception but would have created another problem
because size of the arrays would have to be guessed. Therefore the fix
was to break backward compatibility and changing method's signature to
return an array. This assumes the method is not widely used yet but it
would not be useful anyway for dynamic parameters.