Fix PVT not working because of wrong Thread
PVT used a new Thread for executing a Runnable which should also modify
Views. This caused a CalledFromWrongThreadException because a View must
only be modified from Activity's UI thread. As this new Thread was only
used to execute the Runnable once this was fixed by simply replacing it
with posting it to the Handler which already executes remaining anyway.
The synchronized statement in run() was kept although it is unnecessary
to allow easier refactoring later.