Fix PVT ignoring parameters
PVT ignored the two parameters passed with starting Intent and therefore
only used the default parameters. The fault was method fillInt() of base
class CognitiveActivity. The method actually read parameters but did not
return them to the caller. It was written as it would be possible to set
variables outside the method by assigning input parameters. In Java this
is not possible. Therefore the read parameters were never forwarded. The
problem was fixed by adding a return type to the method. Input parameter
is used as a default value if no parameter was found. The fix is somehow
backward compatible if everything is compiled again. Because of modified
method signatures it will not work if linking dynamically. But it may be
ignored here. The input type was therefore also changed to avoid boxing.

Also fixed same problem in locally unused fillBoolean() for consistency.