Fix PVT using the wrong default parameters
PVT used static variables for keeping input parameters. This may work in
most cases but is actually unpredictable. It is because static variables
are kept as long as an application process is running. Starting Activity
again means that the old values are reused. Starting an Activity without
input parameters therefore might not use defaults but some values of the
last start which may be different. Fixed by removing the static keyword.